requirements.txt
flask 
flask_sqlalchemy 
pymysql 
Flask-Migrate 
flask-wtf 
